requirements.txt
python_telegram_bot==21.9
requests==2.32.3
pytz==2024.2
PyMySQL==1.1.1
urllib3==2.2.3
cryptography==44.0.0